Added pause and ball/wall collision detection
Added a pause function triggered by the pause key on the
keyboard.

Added collision detection with all four walls. The ball will be
reflected back at the same angle if it collides with a wall.
The collision detection with the left and right walls will eventually
be used to control the score.

There is no bat/ball collision detection at this stage.